Fixed hanging run in split_explicit.
Run was hanging during timer_write because of threads with different
timer information.

The MPAS timer funcationality is only used on the master process now,
but each thread calls timer_start and stop to call the tau start and
stop. This way profiling can be done on multi-threaded applications
without using the MPAS timers.